pyproject.toml
[project]
name = "autoresearch"
version = "0.1.0"
description = "Autonomous pretraining research swarm"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "requests>=2.32.0",
    "vectorbt>=0.26.0",
    "yfinance>=0.2.40",
    "ccxt>=4.2.0",
    "arxiv>=2.1.0",
    "bm25s>=0.1.0",
    "pypdf>=4.0.0",
    "scipy>=1.13.0",
    "seaborn>=0.13.2",
    "fastparquet>=2024.2.0",
    "groq>=0.4.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.5.0",
    "typer>=0.9.0",
    "wandb>=0.16.0",
    "openai>=2.26.0",
]


